# api/openapi.yaml
openapi: 3.0.3
info:
  title: Enzyme API
  description: REST API for Enzyme, a self-hostable Slack alternative
  version: 1.0.0
servers:
  - url: /api
    description: API server
...
paths:
...
  /workspaces/create:
    post:
      tags: [workspaces]
      summary: Create a new workspace
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceInput'
      responses:
        '200':
          description: Workspace created
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /workspaces/{wid}:
    get:
      tags: [workspaces]
      summary: Get workspace details
      operationId: getWorkspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{wid}/update:
    post:
      tags: [workspaces]
      summary: Update workspace
      operationId: updateWorkspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceInput'
      responses:
        '200':
          description: Workspace updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{wid}/members/list:
    post:
      tags: [workspaces]
      summary: List workspace members
      operationId: listWorkspaceMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceMemberWithUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{wid}/members/remove:
    post:
      tags: [workspaces]
      summary: Remove a member from workspace
      operationId: removeWorkspaceMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{wid}/members/update-role:
    post:
      tags: [workspaces]
      summary: Update member role
      operationId: updateWorkspaceMemberRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role]
              properties:
                user_id:
                  type: string
                role:
                  $ref: '#/components/schemas/WorkspaceRole'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{wid}/invites/create:
    post:
      tags: [workspaces]
      summary: Create an invite
      operationId: createWorkspaceInvite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteInput'
      responses:
        '200':
          description: Invite created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite:
                    $ref: '#/components/schemas/Invite'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workspaces/notifications:
    get:
      tags: [workspaces]
      summary: Get notification summaries for all workspaces
      operationId: getWorkspaceNotifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification summaries per workspace
          content:
            application/json:
              schema:
                type: object
                required: [workspaces]
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkspaceNotificationSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /workspaces/reorder:
    post:
      tags: [workspaces]
      summary: Reorder workspaces for current user
      operationId: reorderWorkspaces
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderWorkspacesInput'
      responses:
        '200':
          description: Workspaces reordered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /invites/{code}/accept:
    post:
      tags: [workspaces]
      summary: Accept an invite
      operationId: acceptInvite
      security:
        - bearerAuth: []
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invite accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Channel endpoints
...
  /channels/{id}/members/add:
    post:
      tags: [channels]
      summary: Add member to channel
      operationId: addChannelMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/channelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                role:
                  $ref: '#/components/schemas/ChannelRole'
      responses:
        '200':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /workspaces/{wid}/unreads:
    post:
      tags: [messages]
      summary: List all unread messages across channels
      operationId: listAllUnreads
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  default: 50
                cursor:
                  type: string
      responses:
        '200':
          description: List of unread messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadMessagesResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /workspaces/{wid}/icon:
    post:
      tags: [workspaces]
      summary: Upload workspace icon
      operationId: uploadWorkspaceIcon
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Icon uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceIconUploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
    delete:
      tags: [workspaces]
      summary: Remove workspace icon
      operationId: deleteWorkspaceIcon
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
      responses:
        '200':
          description: Icon removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # SSE endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    workspaceId:
      name: wid
      in: path
      required: true
      schema:
        type: string
      description: Workspace ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    Workspace:
      type: object
      required: [id, name, settings, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        icon_url:
          type: string
        settings:
          type: string
          description: JSON string containing workspace settings (for backward compatibility)
        parsed_settings:
          $ref: '#/components/schemas/WorkspaceSettings'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, member, guest]
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    CreateWorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateWorkspaceInput:
      type: object
      properties:
        name:
          type: string
        settings:
          $ref: '#/components/schemas/WorkspaceSettings'
...